refactor: extract shared hooks/utils, separate canvas module, parallelize ProcessPaths
- Extract useToast hook to eliminate duplicate toast logic between App.tsx and SettingsWindow.tsx
- Create utils.ts with getExportInfo() and getQualityFromBPP() to consolidate repeated filename/MIME handling
- Move renderImageToCanvas (150 lines) into canvas.ts with typed RenderSettings interface
- Unify handleAddFiles/handleAddFolder into generic handleOpenImages (remove unused handleAddFolder)
- Unify setFrameColor/setTextColor into generic setPerImageColor helper
- Optimize image loading useEffect deps to avoid full-array re-evaluation on each load
- Parallelize Go ProcessPaths with bounded goroutine pool (NumCPU workers)

App.tsx: 1111 → 882 lines (-21%)
